pyproject.toml
[tool.poetry]
name = "latoken-khakaton-tg-bot"
version = "100.99.0"
description = "Chat bot"
authors = ["Fakhriddin <fahriddin3040@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
python-telegram-bot = "^21.10"
python-dotenv = "^1.0.1"
django = "^5.1.4"
openai = "1.0.0"
asgiref = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
